Move codegen to an isolated directory
This moves all the Java codegen tooling into its own directory.
This is being done because this repo will also house the base
Go components of the Smithy client, and from a Go perspective it
makes more sense to have the root be a go package. Go customers
looking at the repo will be less confused.