Add symbol formatter that emits pointers
There are cases where a single symbol might need to be referenced as
either a pointer or a concrete value. Notably structures will need to
be referenced as a pointer in member values and as a concrete value
when actually defining their struct types.

This also updates structures to be properly referenced as pointers in
member values.