Run go fmt before generating go.mod
go fmt tries to fetch packages in the go.mod file. This is problematic
because we may introduce dependencies on new features which will require
a version bump of the smithy-go package. To avoid this, we format the
code before generating the go.mod file. We know the formatting will be
correct anyway, as it's being generated by the go tooling.