Update middleware Metadata to be a struct instead of interface
Updates the middleware Metadata type to be pass by value as a structure
in the middleware instead of interface. This removes the need from
middleware writers to use `NewMetadata` each time an error is handled.
This does require the helper methods to use the Metadata pointer for
set, and MetadataReader for Gets.