Move types into their own namespace
This moves generated types for most shapes into a "types" namespace.
This allows for things like synthesized types for operation input /
output without risk of conflicting names.

As part of this a symbol is generated for the service shape (client)
to enable generating an empty file for the client. This is needed as
otherwise `go fmt` will complain about an empty package.