Update namespace handling
This updates and fixes the way namespacing is handled, particularly
with regard to aliasing. Previously a SymbolReference was being
created to represent the namespace, but that isn't right since a
namespace isn't an individually referenceable entity. Now the symbol
itself controls how its namespace should be referenced. This ends up
being much more concise to define as an added bonus.