Use version ranges for Smithy dependencies
This updates the dependencies on smithy to use a range across the
current minor version. This will let us pick up any new patch
versions without having to manually update every time.

Range dependencies can be very dangerous, but in this case it's only
on a dependency that we control and the range is fairly tight.