Fix Insert After relative ordering, added unit tests
Adds unit tests for the middleware ordering behavior, this discovered a
bug in the Insert After behavior not correctly inserting within a list.
It only inserted after to the end of the list.

Also adds a List method for all middleware step that returns the list of
middleware in the step.

Reduce allocations for getting order of middleware when middleware stack
is built into handler decorators.