Don't use test server for response tests
This removes the test server for response tests, replacing it with
a simple stubbed HTTP client that returns a pre-canned response.
This is necessary because the test server would inject additional
headers / info that we weren't expecting which would break protocol
tests where we have an expectation on the entire set of headers.