Support endpoint and hostLabel traits
This adds support for the endpoint and hostLabel traits. This is
accomplished by adding a `HostPrefix` member to the Smithy Request
struct, which is prepended to the host when the underlying http
Request is finalized.

If there are no labels, the host prefix is simply generated as a
a string that is set on the smithy request like so:

```
request.HostPrefix = "data-"
```

If there are labels, `fmt.Sprintf` is used:

```
request.HostPrefix = fmt.Sprintf("data-%s", *input.ServiceName)
```

While the need of filling out a template is similar to what is done
with the URL path, the restriction of `hostLabel` to string shapes
that are required means that we don't need such a heavy-weight
solution. Since this solution is minimal, the strategy can easily
be changed if needed.

Additionally, this was inlined into the operation serializer for
the sake of simplicity. You necessarily need access to the operation
input struct, so if this were a middleware then it would need to be
generated.