codegen: Add option to build single API, instead of all
Adds support for a new environment variable SMITHY_GO_BUILD_API that
allows you to specify the serviceId of the API model that should be
built.

The projects for the models will still be invoked since that is done in
smithy core, but the smithy-go build of the model will not be done.

Example using aws-sdk-go-v2 to build a single API client.

  SMITHY_GO_BUILD_API="com.amazonaws.ec2#AmazonEC2" make smithy-build