codegen: Introduce the concept of a NullableIndex and Pointable Index (#232)
Adds a new KnowedgeIndex implementation GoPointableIndex this index declares when a shape should be a pointer or value type in Go code. Updates the SDK's code generation so that all non-sparse list and maps are rendered with value members instead of pointers. This null-ability applies to number and boolean members to that are not explicitly decorated as boxed.

Fixes #128
Fixes #127
Fixes #149