codegen: Update generated serde to skip unmodeled input/outputs
Updates the SDK's generated serializers and deserializers to skip unmodeled
input and output shapes for operations.

This issue caused generated deserializers to attempt to deserialize
operation responses that were not modeled, resulting in an error.
Whereas the SDK should of ignored the response if one was present for
operations without a modeled output.

Related to https://github.com/aws/aws-sdk-go-v2/issues/1047