codegen: Update generated serde to skip unmodeled input/outputs (#261)
Updates the SDK's generated serializers and deserializers to skip unmodeled
input and output shapes for operations.

This issue caused generated deserializers to attempt to deserialize
operation responses that were not modeled, resulting in an error.
Whereas the SDK should of ignored the response if one was present for
operations without a modeled output.

Related to https://github.com/aws/aws-sdk-go-v2/issues/1047
Fixes #262

#### Input
Unmodeled Input shape serializer functions are not generated.

This update changes API protocol operation serializers for HTTP binding protocols to skip serialization of the request document if the operation's input was not modeled.

This change is **not** applied to HTTP RPC protocol serializations. This is due to the AwsQuery protocol requiring some protocol specific elements to always be serialized regardless if an input is modeled or not.

#### Output
Unmodeled output shape serializer functions are not generated.

This update changes all API protocol operation deserializers to discard the response body if the operation's output was not modeled.