Add method to get successful output from waiters (#319)
This PR adds a WaitForOutput method to waiters that returns the successful operation's output. This allows the user to avoid making an extra API request to find the full result of the operation that they were waiting for - in my case, I want to access the StandardOutputContent on the GetCommandInvocationOutput using the ssm.NewCommandExecutedWaiter.

I opted for adding a method rather than changing the existing Wait method to avoid breaking the interface, but if a breaking change is ok then I'm happy to update this to just change the signature of Wait to include the output.

I've regenerated the SDK code locally and confirmed that the new method works as intended. The generated code changes from 697d3ed can be seen in haines/aws-sdk-go-v2@2ed786e

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.