[issues-391] Fix document typecheck in encoding  for nested types
This resolves github issue 391 where encoding of document types
is not checking to see if nested types are explicitly marked
for no serialization or deserialization. In order to to check
for every layer of nesting, I moved the noserde check to the
recursive layer from the original top-level function. This
mimics how it correctly works in decoding