Update intEnum generation to use type alias
Due to type definitions requiring strict casting and implementation
overhead in protocol generation, this commit uses type aliases to
generate intEnums as an intermediate step until migrating back to
type definitions.

This change is backwards compatible except for the removed `Values()`
method that was on intEnum type definitions, but not definable on type
aliases since defining methods on base types (`int32`) is not permitted.