Add `ErrorCodeOverride` to error structures
Adding the `ErrorCodeOverride` field allows protocols to
override the default `ErrorCode()` implementation (which
is the name of the error) with a custom string.

If a smithy model uses the `ErrorCodeOverride` field in
an error structure, it will be renamed to
`ErrorCodeOverride_`.

Also, the writer for the switch-statement default block
is now pluggable, which means the default block can be
overwritten with a custom writer.