Fix time parsing for non-GMT timezone
Replace httpDateFormat with time.RFC1123 in ParseHTTPDate() for parsing
responses not using GMT. The stdlib time package uses MST for the
timezone in the format strings. Having GMT in the format string is only
valid if the time string to be parsed is also GMT.

This was causing the following parse errors for a HeadObject
Last-Modified time value when response was in UTC:
parsing time "Thu, 01 Jan 1970 00:00:00 UTC"
as "Mon, 02 Jan 2006 15:04:05 GMT":
cannot parse "UTC" as " GMT"#012

This also adds test cases for UTC date strings.

httpDateFormat renamed to httpDateFormatOutputGMT to clarify this is for
output formatting compatible with RFC requiring GMT timezone. This
should not be used to parse input strings.