Add Go Synthetic Trait equality methods
When using smithy diff or trait comparisons for the Go `Synthetic`
trait, the following exception is thrown in the `createNode()`
implementation:

```text
"attempted to serialize runtime only trait"
```

This is due to inheriting from `AbstractTrait`'s equality methods which
use the `ToNode()` method which calls the `createNode()` method.

This change adds an `equals()` and `hashCode()` implementation for the
Go `Synthetic` trait that does not use `ToNode()` or `createNode()`.

`ToNode()` cannot be overriden since `AbstractTrait`'s implementation is
marked as `final`.